seed-data/requirements.txt
# DPO Dataset Generation Requirements

# Core dependencies
hydra-core>=1.3.0
omegaconf>=2.3.0
pandas>=2.0.0
openai>=1.0.0
ollama>=0.1.0

# Utilities
python-dotenv>=1.0.0

# Development
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0
